+++ src/test/java/org/folio/des/scheduling/acquisition/AcqBaseExportTaskTriggerTest.java
@@ -1,5 +1,6 @@
 package org.folio.des.scheduling.acquisition;
 
+import static java.util.Map.entry;
 import static org.folio.des.domain.dto.ScheduleParameters.SchedulePeriodEnum;
 import static org.folio.des.domain.dto.ScheduleParameters.WeekDaysEnum;
 
@@ -147,19 +148,23 @@ class AcqBaseExportTaskTriggerTest {
     long diff = TimeUnit.HOURS.convert(diffInMilliseconds, TimeUnit.MILLISECONDS);
     assertEquals(expDiffHours, diff);
 
-    ZonedDateTime nowDateTime = Instant.now().atZone(ZoneId.of("UTC")).truncatedTo(ChronoUnit.DAYS);
-    Instant lastInstant = Instant.ofEpochMilli(actDateHourLate.getTime()).truncatedTo(ChronoUnit.DAYS);
+    ZonedDateTime expDateTime =
+      Instant.now().atZone(ZoneId.of("UTC")).plusHours(expDiffHours).truncatedTo(ChronoUnit.DAYS);
+    Instant lastInstant = Instant.ofEpochMilli(actDate.getTime()).truncatedTo(ChronoUnit.DAYS);
     ZonedDateTime zonedDateTime = ZonedDateTime.ofInstant(lastInstant, ZoneId.of("UTC")).truncatedTo(ChronoUnit.DAYS);
-    assertTrue(nowDateTime.isEqual(zonedDateTime));
+    assertTrue(expDateTime.isEqual(zonedDateTime));
   }
 
   @Test
   @DisplayName("Should increase Days and provided time should be changed")
   void dailySchedule() {
     int expDiffDays = 3;
+    ZoneId zoneId = ZoneId.of("Asia/Shanghai");
+    ZonedDateTime nowTime = getNowTime(zoneId);
+    ZonedDateTime scheduledDateTime = nowTime.plusHours(1);
     ScheduleParameters scheduleParameters = new ScheduleParameters();
     scheduleParameters.setId(UUID.randomUUID());
-    scheduleParameters.setScheduleTime("11:12:13");
+    scheduleParameters.setScheduleTime(scheduledDateTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
     scheduleParameters.setScheduleFrequency(expDiffDays);
     scheduleParameters.setSchedulePeriod(SchedulePeriodEnum.DAY);
     scheduleParameters.setTimeZone(ASIA_SHANGHAI_ZONE);
@@ -170,9 +175,9 @@ class AcqBaseExportTaskTriggerTest {
 
     Instant firstInstant = Instant.ofEpochMilli(actDate.getTime());
     ZonedDateTime firstZonedDateTime = ZonedDateTime.ofInstant(firstInstant, ZoneId.of(ASIA_SHANGHAI_ZONE));
-    String firstDateStr = firstZonedDateTime.toString();
-    int dayOfMonth = firstZonedDateTime.getDayOfMonth();
-    assertTrue(firstDateStr.contains(dayOfMonth+"T11:12:13+08:00[Asia/Shanghai]"));
+    assertEquals(scheduledDateTime.getDayOfMonth(), firstZonedDateTime.getDayOfMonth());
+    assertEquals(scheduledDateTime.getHour(), firstZonedDateTime.getHour());
+    assertEquals(scheduledDateTime.getMinute(), firstZonedDateTime.getMinute());
 
     //Second try
     triggerContext.update(actDate, actDate, actDate);
@@ -181,11 +186,12 @@ class AcqBaseExportTaskTriggerTest {
     long diff = TimeUnit.DAYS.convert(diffInMilliseconds, TimeUnit.MILLISECONDS);
     assertEquals(expDiffDays, diff);
 
-    Instant instant = Instant.ofEpochMilli(actDateDayLate.getTime());
-    ZonedDateTime secondZonedDateTime = ZonedDateTime.ofInstant(instant, ZoneId.of(ASIA_SHANGHAI_ZONE));
-    dayOfMonth = firstZonedDateTime.plusDays(expDiffDays).getDayOfMonth();
-    String lastDateStr = secondZonedDateTime.toString();
-    assertTrue(lastDateStr.contains(dayOfMonth+"T11:12:13+08:00[Asia/Shanghai]"));
+    Instant secondInstant = Instant.ofEpochMilli(actDateDayLate.getTime());
+    ZonedDateTime secondZonedDateTime = ZonedDateTime.ofInstant(secondInstant, zoneId);
+    assertEquals(scheduledDateTime.plusDays(expDiffDays).getDayOfMonth(), secondZonedDateTime.getDayOfMonth());
+    assertEquals(scheduledDateTime.getHour(), secondZonedDateTime.getHour());
+    assertEquals(scheduledDateTime.getMinute(), secondZonedDateTime.getMinute());
+
     //Third try
     triggerContext.update(actDateDayLate, actDateDayLate, actDateDayLate);
     final Date thirdDateDayLate = trigger.nextExecutionTime(triggerContext);
@@ -195,9 +201,32 @@ class AcqBaseExportTaskTriggerTest {
 
     Instant thirdInstant = Instant.ofEpochMilli(thirdDateDayLate.getTime());
     ZonedDateTime thirdZonedDateTime = ZonedDateTime.ofInstant(thirdInstant, ZoneId.of(ASIA_SHANGHAI_ZONE));
-    dayOfMonth = secondZonedDateTime.plusDays(expDiffDays).getDayOfMonth();
-    String thirdLastDateStr = thirdZonedDateTime.toString();
-    assertTrue(thirdLastDateStr.contains(dayOfMonth+"T11:12:13+08:00[Asia/Shanghai]"));
+    assertEquals(scheduledDateTime.plusDays(expDiffDays * 2).getDayOfMonth(), thirdZonedDateTime.getDayOfMonth());
+    assertEquals(scheduledDateTime.getHour(), thirdZonedDateTime.getHour());
+    assertEquals(scheduledDateTime.getMinute(), thirdZonedDateTime.getMinute());
+  }
+
+  @Test
+  @DisplayName("Daily run should be executed tomorrow in case when scheduled datetime less than current datetime")
+  public void dailyScheduleWhenScheduledDateLessThanCurrentDate() {
+    //Given
+    ZonedDateTime nowTime = getNowTime();
+    ZonedDateTime scheduledDateTime = nowTime.minusHours(1);
+    ScheduleParameters scheduleParameters = new ScheduleParameters();
+    scheduleParameters.setScheduleTime(scheduledDateTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
+    scheduleParameters.setScheduleFrequency(1);
+    scheduleParameters.setSchedulePeriod(SchedulePeriodEnum.DAY);
+    AcqBaseExportTaskTrigger trigger = new AcqBaseExportTaskTrigger(scheduleParameters, null, true);
+
+    //When
+    SimpleTriggerContext triggerContext = new SimpleTriggerContext();
+    final Date actDate = trigger.nextExecutionTime(triggerContext);
+
+    //Then
+    ZonedDateTime actDateTime = getActualTime(actDate);
+    assertEquals(scheduledDateTime.plusDays(1).getDayOfMonth(), actDateTime.getDayOfMonth());
+    assertEquals(scheduledDateTime.getHour(), actDateTime.getHour());
+    assertEquals(scheduledDateTime.getMinute(), actDateTime.getMinute());
   }
 
   @DisplayName("Weekly job scheduled for specific hour, when last time behind current time")
@@ -405,10 +434,14 @@ class AcqBaseExportTaskTriggerTest {
   }
 
   private ZonedDateTime getNowTime() {
+    return getNowTime(ZoneId.of("UTC"));
+  }
+
+  private ZonedDateTime getNowTime(ZoneId zoneId) {
     Calendar scheduledCal = Calendar.getInstance();
     Date scheduledDate = scheduledCal.getTime();
     Instant scheduledInstant = Instant.ofEpochMilli(scheduledDate.getTime());
-    return ZonedDateTime.ofInstant(scheduledInstant, ZoneId.of("UTC"));
+    return ZonedDateTime.ofInstant(scheduledInstant, zoneId);
   }
 
   private ZonedDateTime getActualTime(Date actDate) {
@@ -419,14 +452,14 @@ class AcqBaseExportTaskTriggerTest {
   @DisplayName("Hour job scheduled for specific hour, when last time behind current time")
   @ParameterizedTest
   @CsvSource({
-   "1, 3, 1",
-   "2, 3, 1",
-   "4, 3, 1",
-   "5, 3, 2",
-   "6, 3, 3",
-   "7, 3, 4",
-   "3, 0, 3",
-   "3, 3, 1"
+    "1, 3, 1",
+    "2, 3, 1",
+    "4, 3, 1",
+    "5, 3, 2",
+    "6, 3, 3",
+    "7, 3, 4",
+    "3, 0, 3",
+    "3, 3, 1"
   })
   void hourlyScheduleWithHours(int frequency, int addHours, int expDiffHours) {
     ScheduleParameters scheduleParameters = new ScheduleParameters();
@@ -451,7 +484,7 @@ class AcqBaseExportTaskTriggerTest {
     Instant actInstant = Instant.ofEpochMilli(actDate.getTime());
     ZonedDateTime actZonedDateTime = ZonedDateTime.ofInstant(actInstant, ZoneId.of("UTC"));
     int actHour = actZonedDateTime.getHour();
-    assertEquals(actHour, nowHour + expDiffHours + addHours);
+    assertEquals((nowHour + expDiffHours + addHours) % 24, actHour);
   }
 
   @DisplayName("Hour job scheduled for specific hour, when last time behind current time")
@@ -489,6 +522,6 @@ class AcqBaseExportTaskTriggerTest {
     Instant actInstant = Instant.ofEpochMilli(actDate.getTime());
     ZonedDateTime actZonedDateTime = ZonedDateTime.ofInstant(actInstant, ZoneId.of("UTC"));
     int actHour = actZonedDateTime.getHour();
-    assertEquals(nowHour + expDiffHours + addHours, actHour);
+    assertEquals((nowHour + expDiffHours + addHours) % 24, actHour);
   }
 }